+++ src/utils/deser.rs
@@ -4,30 +4,31 @@ use tracing::debug;
 
 use crate::{codec::consts::keys, errors::DecodingError};
 
-fn value_to_map(value: Value) -> Result<Vec<(Value, Value)>, DecodingError> {
+pub fn value_to_map(value: Value) -> Result<Vec<(Value, Value)>, DecodingError> {
     match value {
         Value::Map(x) => Ok(x),
         rest => Err(DecodingError::type_mismatch("map", rest.to_string())),
     }
 }
 
-/// Extract IPROTO_DATA from response body and deserialize it.
-pub(crate) fn extract_iproto_data(value: Value) -> Result<Value, DecodingError> {
-    let map = value_to_map(value).map_err(|err| err.in_other("OK response body"))?;
+pub(crate) fn find_and_take_single_key_in_map(key: u8, map: Vec<(Value, Value)>) -> Option<Value> {
     for (k, v) in map {
-        if matches!(k, Value::Integer(x) if x.as_u64().map_or(false, |y| y == keys::DATA as u64)) {
-            return Ok(v);
-        } else {
-            // NOTE: no errors or warnings in case protocol adds new data in responsese in future
-            // TODO: configurable logging level?
-            debug!("Unexpected key encountered in response body: {:?}", k);
+        if matches!(k, Value::Integer(x) if x.as_u64().map_or(false, |y| y == key as u64)) {
+            return Some(v);
         }
     }
-    Err(DecodingError::missing_key("DATA").in_other("OK response body"))
+    None
+}
+
+/// Extract IPROTO_DATA from response body and deserialize it.
+pub fn extract_iproto_data(value: Value) -> Result<Value, DecodingError> {
+    let map = value_to_map(value).map_err(|err| err.in_other("OK response body"))?;
+    find_and_take_single_key_in_map(keys::DATA, map)
+        .ok_or_else(|| DecodingError::missing_key("DATA").in_other("OK response body"))
 }
 
 /// Extract IPROTO_DATA from response body and deserialize it into provided type.
-pub(crate) fn extract_and_deserialize_iproto_data<T: DeserializeOwned>(
+pub fn extract_and_deserialize_iproto_data<T: DeserializeOwned>(
     value: Value,
 ) -> Result<T, DecodingError> {
     extract_iproto_data(value).and_then(|x| rmpv::ext::from_value(x).map_err(Into::into))