+++ src/utils/mod.rs
@@ -1,7 +1,9 @@
 //! Various helpers.
 
+pub use self::deser::{extract_and_deserialize_iproto_data, extract_iproto_data, value_to_map};
+
 pub(crate) use self::{
-    deser::{extract_and_deserialize_iproto_data, extract_iproto_data},
+    deser::find_and_take_single_key_in_map,
     unique_id_name_map::{UniqueIdName, UniqueIdNameMap},
 };